improve code climate issues
