+++ .prettierrc
@@ -2,5 +2,6 @@
     "trailingComma": "all",
     "tabWidth": 2,
     "semi": true,
-    "singleQuote": false
+    "singleQuote": false,
+    "printWidth": 120,
 }
