+++ src/components/contracts/ContractStatus.js
@@ -13,10 +13,7 @@ import CheckCircleIcon from "@material-ui/icons/CheckCircle";
 import CancelIcon from "@material-ui/icons/Cancel";
 import InfoIcon from "@material-ui/icons/Info";
 
-const getOverlaps = (contracts) =>
-  contracts.map((c) => (
-    <div key={c.id}>{`${c.orgUnit.name}, id: ${c.id}`}</div>
-  ));
+const getOverlaps = (contracts) => contracts.map((c) => <div key={c.id}>{`${c.orgUnit.name}, id: ${c.id}`}</div>);
 
 const styles = (theme) => ({
   ...tablesStyles(theme),
@@ -46,25 +43,12 @@ const ContractStatus = ({ contract, t }) => {
   };
   const openPopOver = Boolean(anchorEl);
   const id = openPopOver ? "contract-status" : undefined;
-  const {
-    coverageIssue,
-    nonVisibleOverlaps,
-    visibleOverlaps,
-  } = contract.statusDetail;
+  const { coverageIssue, nonVisibleOverlaps, visibleOverlaps } = contract.statusDetail;
   return contract.status ? (
-    <CheckCircleIcon
-      className={classNames(classes.success, classes.tableIcon)}
-    />
+    <CheckCircleIcon className={classNames(classes.success, classes.tableIcon)} />
   ) : (
     <>
-      <CancelIcon
-        className={classNames(
-          classes.error,
-          classes.tableIcon,
-          classes.pointer,
-        )}
-        onClick={handleClick}
-      />
+      <CancelIcon className={classNames(classes.error, classes.tableIcon, classes.pointer)} onClick={handleClick} />
       <Popover
         id={id}
         open={openPopOver}
@@ -89,12 +73,7 @@ const ContractStatus = ({ contract, t }) => {
 
               <Tooltip arrow title={getOverlaps(visibleOverlaps)}>
                 <span>
-                  <IconButton
-                    size="small"
-                    disableRipple
-                    disableFocusRipple
-                    className={classes.button}
-                  >
+                  <IconButton size="small" disableRipple disableFocusRipple className={classes.button}>
                     <InfoIcon color="action" />
                   </IconButton>
                 </span>
@@ -108,12 +87,7 @@ const ContractStatus = ({ contract, t }) => {
               })}
               <Tooltip arrow title={getOverlaps(nonVisibleOverlaps)}>
                 <span>
-                  <IconButton
-                    size="small"
-                    disableRipple
-                    disableFocusRipple
-                    className={classes.button}
-                  >
+                  <IconButton size="small" disableRipple disableFocusRipple className={classes.button}>
                     <InfoIcon color="action" />
                   </IconButton>
                 </span>
